requirements.txt
# Core:
Django==2.2.7
#MySQL-python==1.2.5
django-grappelli==2.8.1
psycopg2
Pillow==5.2.0
django-cors-headers==3.2.1
openpyxl==2.3.5
pytz==2016.4
python-dateutil==2.5.3
djangorestframework==3.10.3
django-smart-selects==1.5.4
django-rest-swagger==2.2.0
djangorestframework-simplejwt==4.3.0
djangorestframework-jwt==1.11.0
python-social-auth==0.3.6
social-auth-app-django==3.1.0
djangorestframework-link-header-pagination==0.1.1
django-environ==0.4.5
redis==3.4.1
requests==2.22.0

#culqi


#django-smart-selects==1.5.4


# Docs:
# git+https://github.com/sphinx-doc/sphinx@stable
# sphinxcontrib-fulltoc

# Required by commons and to build the docs:
# django_facebook
# PySimpleSOAP==1.16
# requests

# Production:
# gunicorn
# supervisor